+++ src/TF1.cpp
@@ -35,6 +35,8 @@ struct TF : Module  {
 		LIGHT_BG_RED,
 		LIGHT_BG_GREEN,
 		LIGHT_BG_BLUE,
+		LIGHT_LEFT,
+		LIGHT_RIGHT,
 		NUM_LIGHTS
 	};
 
@@ -62,6 +64,8 @@ struct TF : Module  {
 		configOutput(OUTPUT_TOR, "Deprecated");
 		configLight(LIGHT_FG_RED, "Foreground Colour");
 		configLight(LIGHT_BG_RED, "Background Colour");
+		configLight(LIGHT_LEFT, "Module Link");
+		configLight(LIGHT_RIGHT, "Module Link");
 		prevValues[0] = 0.1569f;
 		prevValues[1] = 0.6902f;
 		prevValues[2] = 0.9529f;
@@ -118,6 +122,8 @@ struct TF : Module  {
 		}
 		leftExpander.messageFlipRequested = true;
 		rightExpander.messageFlipRequested = true;
+		lights[LIGHT_LEFT].setBrightness(leftExpander.module && ((leftExpander.module->model == modelTD202) || (leftExpander.module->model == modelTD316)));
+		lights[LIGHT_RIGHT].setBrightness(rightExpander.module && ((rightExpander.module->model == modelTD202) || (rightExpander.module->model == modelTD316)));
 	}
 	std::string encodeColor(float r, float g, float b) {
 		std::string out;
@@ -166,6 +172,9 @@ struct TF101 : SchemeModuleWidget {
 
 		addChild(createLightCentered<MediumLight<RGBLight>>(Vec(14.5, 55.5), module, TF<true>::LIGHT_FG_RED));
 		addChild(createLightCentered<MediumLight<RGBLight>>(Vec(14.5, 189.5), module, TF<true>::LIGHT_BG_RED));
+		
+		addChild(createLightCentered<LeftLight>(Vec(3, 14), module, TF<true>::LIGHT_LEFT));
+		addChild(createLightCentered<RightLight>(Vec(87, 14), module, TF<true>::LIGHT_RIGHT));
 
 		addOutput(createOutputCentered<DeprecatedPort>(Vec(73.5,31.5), module, TF<true>::OUTPUT_TOR));
 	}
@@ -209,10 +218,13 @@ struct TF102 : SchemeModuleWidget {
 		addInput(createInputCentered<SilverPort>(Vec(15,265), module, TF<false>::INPUT_BG_BLUE));
 		addInput(createInputCentered<SilverPort>(Vec(15,315), module, TF<false>::INPUT_FONT_SIZE));
 
-		addChild(createLightCentered<MediumLight<RGBLight>>(Vec(15, 71), module, TF<true>::LIGHT_FG_RED));
-		addChild(createLightCentered<MediumLight<RGBLight>>(Vec(15, 186), module, TF<true>::LIGHT_BG_RED));
+		addChild(createLightCentered<MediumLight<RGBLight>>(Vec(15, 71), module, TF<false>::LIGHT_FG_RED));
+		addChild(createLightCentered<MediumLight<RGBLight>>(Vec(15, 186), module, TF<false>::LIGHT_BG_RED));
 
-		addOutput(createOutputCentered<DeprecatedPort>(Vec(15,40), module, TF<true>::OUTPUT_TOR));
+		addChild(createLightCentered<LeftLight>(Vec(3, 14), module, TF<false>::LIGHT_LEFT));
+		addChild(createLightCentered<RightLight>(Vec(27, 14), module, TF<false>::LIGHT_RIGHT));
+
+		addOutput(createOutputCentered<DeprecatedPort>(Vec(15,40), module, TF<false>::OUTPUT_TOR));
 	}
 	void render(NVGcontext *vg, SchemeCanvasWidget *canvas) override {
 		drawBase(vg, "TF-102");